# docs/swagger.json
{
    "produces": [
        "application/json"
    ],
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "This is a REST API server for threat registry management.\nIt provides endpoints for managing products, instances, threats, controls, domains, and tags.",
        "title": "Threatreg API",
        "contact": {
            "name": "API Support",
            "email": "support@threatreg.io"
        },
        "license": {
            "name": "MIT",
            "url": "https://opensource.org/licenses/MIT"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/api/v1",
    "paths": {
...
        "/threats": {
            "get": {
                "description": "Get a list of all security threats in the system",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Threats"
                ],
                "summary": "List all threats",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handlers.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.Threat"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "description": "Create a new security threat with the provided title and description",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Threats"
                ],
                "summary": "Create a new threat",
                "parameters": [
                    {
                        "description": "Threat creation request",
                        "name": "threat",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.CreateThreatRequest"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handlers.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.Threat"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/threats/by-domain/{domainId}/unresolved": {
            "get": {
                "description": "Get threats in a domain showing count of instances that haven't resolved each threat",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Threats"
                ],
                "summary": "List threats by domain with unresolved instance counts",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Domain ID (UUID)",
                        "name": "domainId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handlers.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/models.ThreatWithUnresolvedByInstancesCount"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/threats/{id}": {
            "get": {
                "description": "Get a specific security threat by its UUID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Threats"
                ],
                "summary": "Get a threat by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Threat ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handlers.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.Threat"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "description": "Update a threat's title and/or description",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Threats"
                ],
                "summary": "Update a threat",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Threat ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Threat update request",
                        "name": "threat",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/handlers.UpdateThreatRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/handlers.SuccessResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/models.Threat"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a security threat by its UUID",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Threats"
                ],
                "summary": "Delete a threat",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Threat ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/handlers.SuccessResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/handlers.ErrorResponse"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
...
        "handlers.CreateThreatRequest": {
            "type": "object",
            "required": [
                "description",
                "title"
            ],
            "properties": {
                "description": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "handlers.ErrorResponse": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "error": {
                    "type": "string"
                },
                "message": {
                    "type": "string"
                }
            }
        },
        "handlers.SuccessResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "message": {
                    "type": "string"
                }
            }
        },
...
        "handlers.UpdateThreatRequest": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "models.Control": {
            "type": "object",
            "properties": {
                "controlAssignments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ControlAssignment"
                    }
                },
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "threatControls": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ThreatControl"
                    }
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "models.ControlAssignment": {
            "type": "object",
            "properties": {
                "control": {
                    "$ref": "#/definitions/models.Control"
                },
                "controlID": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "threatAssignment": {
                    "$ref": "#/definitions/models.ThreatAssignment"
                },
                "threatAssignmentID": {
                    "type": "integer"
                }
            }
        },
        "models.Domain": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "instances": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Instance"
                    }
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "models.Instance": {
            "type": "object",
            "properties": {
                "domains": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Domain"
                    }
                },
                "id": {
                    "type": "string"
                },
                "instanceOf": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "product": {
                    "$ref": "#/definitions/models.Product"
                },
                "threatAssignments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ThreatAssignment"
                    }
                }
            }
        },
...
        "models.Product": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "instances": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.Instance"
                    }
                },
                "name": {
                    "type": "string"
                },
                "threatAssignments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ThreatAssignment"
                    }
                }
            }
        },
...
        "models.Threat": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "threatAssignments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ThreatAssignment"
                    }
                },
                "threatControls": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ThreatControl"
                    }
                },
                "title": {
                    "type": "string"
                }
            }
        },
        "models.ThreatAssignment": {
            "type": "object",
            "properties": {
                "controlAssignments": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ControlAssignment"
                    }
                },
                "id": {
                    "type": "integer"
                },
                "instance": {
                    "$ref": "#/definitions/models.Instance"
                },
                "instanceID": {
                    "type": "string"
                },
                "product": {
                    "$ref": "#/definitions/models.Product"
                },
                "productID": {
                    "type": "string"
                },
                "threat": {
                    "$ref": "#/definitions/models.Threat"
                },
                "threatID": {
                    "type": "string"
                }
            }
        },
        "models.ThreatControl": {
            "type": "object",
            "properties": {
                "control": {
                    "$ref": "#/definitions/models.Control"
                },
                "controlID": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "threat": {
                    "$ref": "#/definitions/models.Threat"
                },
                "threatID": {
                    "type": "string"
                }
            }
        },
        "models.ThreatWithUnresolvedByInstancesCount": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "title": {
                    "type": "string"
                },
                "unresolved_by_instance_count": {
                    "type": "integer"
                }
            }
        }
    },
...
}